brain_monitor/requirements.txt
pandas
dvc
dvc-s3
scipy
python-git
fastapi
uvicorn
python-multipart
